+++ src/main/java/br/com/alura/loja/modelo/Cliente.java
@@ -1,5 +1,6 @@
 package br.com.alura.loja.modelo;
 
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -13,14 +14,18 @@ public class Cliente {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-	private String nome;
-	private String cpf;
+	
+	@Embedded
+	private DadosPessoais dadosPessoais;
 
 	public Cliente() {}
 	
 	public Cliente(String nome, String cpf) {
-		this.nome = nome;
-		this.cpf = cpf;
+		this.dadosPessoais = new DadosPessoais(nome,cpf);
+	}
+	
+	public String getNome() {
+		return this.dadosPessoais.getNome();
 	}
 	
 	public Long getId() {
@@ -31,19 +36,7 @@ public class Cliente {
 		this.id = id;
 	}
 
-	public String getNome() {
-		return nome;
-	}
-
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
-
-	public String getCpf() {
-		return cpf;
-	}
-
-	public void setCpf(String cpf) {
-		this.cpf = cpf;
+	public DadosPessoais getDadosPessoais() {
+		return dadosPessoais;
 	}
 }
