+++ src/main/java/br/com/alura/loja/modelo/DadosPessoais.java
@@ -1,0 +1,25 @@
+package br.com.alura.loja.modelo;
+
+import javax.persistence.Embeddable;
+
+@Embeddable
+public class DadosPessoais {
+
+	private String nome;
+	private String cpf;
+	
+	public DadosPessoais() {}
+	
+	public DadosPessoais(String nome, String cpf) {
+		this.nome = nome;
+		this.cpf = cpf;
+	}
+
+	public String getNome() {
+		return nome;
+	}
+
+	public String getCpf() {
+		return cpf;
+	}
+}
